# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Papers Please System API
  description: API для системы управления УПК (пограничные пункты)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/shifts:
    get:
      tags:
      - shift-controller
      operationId: getAllShifts
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseShiftResponse"
    post:
      tags:
      - shift-controller
      operationId: createShift
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShiftResponse"
  /api/v1/participationsKate:
    get:
      tags:
      - participation-controller
      operationId: getAllParticipations
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseParticipationResponse"
    post:
      tags:
      - participation-controller
      operationId: createParticipation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ParticipationResponse"
  /api/v1/notifications:
    get:
      tags:
      - notification-controller
      operationId: getAllNotifications
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseNotificationResponse"
    post:
      tags:
      - notification-controller
      operationId: createNotification
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResponse"
...
  /api/v1/shifts/{id}:
    get:
      tags:
      - shift-controller
      operationId: getShiftById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShiftResponse"
    delete:
      tags:
      - shift-controller
      operationId: deleteShift
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - shift-controller
      operationId: partialUpdateShift
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShiftResponse"
  /api/v1/participationsKate/{id}:
    get:
      tags:
      - participation-controller
      operationId: getParticipationById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ParticipationResponse"
    delete:
      tags:
      - participation-controller
      operationId: deleteParticipation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - participation-controller
      operationId: partialUpdateParticipation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipationRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ParticipationResponse"
  /api/v1/notifications/{id}:
    get:
      tags:
      - notification-controller
      operationId: getNotificationById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResponse"
    delete:
      tags:
      - notification-controller
      operationId: deleteNotification
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - notification-controller
      operationId: partialUpdateNotification
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationResponse"
...
  /api/v1/participationsKate/by-user/{userId}:
    get:
      tags:
      - participation-controller
      operationId: getParticipationsByUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseParticipationResponse"
  /api/v1/participationsKate/by-specialization/{specialization}:
    get:
      tags:
      - participation-controller
      operationId: getParticipationsBySpecialization
      parameters:
      - name: specialization
        in: path
        required: true
        schema:
          type: string
          enum:
          - PASSPORT
          - LOCALS
          - WORK
          - TRANSIT
          - SPECIAL
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseParticipationResponse"
  /api/v1/participationsKate/by-shift/{shiftId}:
    get:
      tags:
      - participation-controller
      operationId: getParticipationsByShift
      parameters:
      - name: shiftId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseParticipationResponse"
  /api/v1/notifications/by-user/{userId}:
    get:
      tags:
      - notification-controller
      operationId: getNotificationsByUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseNotificationResponse"
  /api/v1/notifications/by-type/{type}:
    get:
      tags:
      - notification-controller
      operationId: getNotificationsByType
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum:
          - TICKET_STATUS_CHANGE
          - TICKET_ASSIGNED
          - SHIFT_INVITATION
          - APPEAL_RESULT
          - EVENT_UPDATE
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseNotificationResponse"
...
components:
  schemas:
...
    ShiftRequest:
      required:
      - createdBy
      - shiftDate
      - upkId
      type: object
      properties:
        shiftDate:
          type: string
          format: date-time
        createdBy:
          type: string
        upkId:
          type: string
    ShiftResponse:
      required:
      - createdBy
      - id
      - shiftDate
      - upkId
      type: object
      properties:
        id:
          type: string
        shiftDate:
          type: string
          format: date-time
        createdBy:
          type: string
        upkId:
          type: string
    ParticipationRequest:
      required:
      - coeffBonus
      - coeffPenalty
      - shiftId
      - specialization
      - userId
      type: object
      properties:
        userId:
          type: string
        shiftId:
          type: string
        coeffBonus:
          type: number
          format: float
        coeffPenalty:
          type: number
          format: float
        specialization:
          type: string
          enum:
          - PASSPORT
          - LOCALS
          - WORK
          - TRANSIT
          - SPECIAL
    ParticipationResponse:
      required:
      - coeffBonus
      - coeffPenalty
      - id
      - shiftId
      - specialization
      - userId
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        shiftId:
          type: string
        coeffBonus:
          type: number
          format: float
        coeffPenalty:
          type: number
          format: float
        specialization:
          type: string
          enum:
          - PASSPORT
          - LOCALS
          - WORK
          - TRANSIT
          - SPECIAL
    NotificationRequest:
      required:
      - message
      - notificationType
      - userId
      type: object
      properties:
        userId:
          type: string
        relatedTicketId:
          type: string
        message:
          type: string
        notificationType:
          type: string
          enum:
          - TICKET_STATUS_CHANGE
          - TICKET_ASSIGNED
          - SHIFT_INVITATION
          - APPEAL_RESULT
          - EVENT_UPDATE
    NotificationResponse:
      required:
      - id
      - isRead
      - message
      - notificationType
      - sentAt
      - userId
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        relatedTicketId:
          type: string
        message:
          type: string
        notificationType:
          type: string
          enum:
          - TICKET_STATUS_CHANGE
          - TICKET_ASSIGNED
          - SHIFT_INVITATION
          - APPEAL_RESULT
          - EVENT_UPDATE
        sentAt:
          type: string
          format: date-time
        isRead:
          type: boolean
...
    ShiftRequestPartial:
      type: object
      properties:
        shiftDate:
          type: string
          format: date-time
        createdBy:
          type: string
        upkId:
          type: string
    ParticipationRequestPartial:
      type: object
      properties:
        shiftId:
          type: string
        userId:
          type: string
        coeffBonus:
          type: number
          format: float
        coeffPenalty:
          type: number
          format: float
        specialization:
          type: string
          enum:
          - PASSPORT
          - LOCALS
          - WORK
          - TRANSIT
          - SPECIAL
    NotificationRequestPartial:
      type: object
      properties:
        userId:
          type: string
        notificationType:
          type: string
          enum:
          - TICKET_STATUS_CHANGE
          - TICKET_ASSIGNED
          - SHIFT_INVITATION
          - APPEAL_RESULT
          - EVENT_UPDATE
        message:
          type: string
        isRead:
          type: boolean
...
    PagedResponseShiftResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShiftResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PagedResponseParticipationResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ParticipationResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PagedResponseNotificationResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NotificationResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
...
    Unit:
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT